+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Timelocks are considered (optionally) in building the `satisfaction` field
 ### Wallet
 
 - Changed `Wallet::{sign, finalize_psbt}` now take a `&mut psbt` rather than consuming it.
+- Require and validate `non_witness_utxo` for SegWit signatures by default, can be adjusted with `SignOptions`
 
 ## [v0.6.0] - [v0.5.1]